requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.7
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
